+++ src/castile/backends/javascript.py
@@ -1,3 +1,4 @@
+from castile.backends.base import BaseCompiler
 from castile.types import Struct, Union
 
 OPS = {
@@ -6,22 +7,8 @@ OPS = {
     '==': '===',
 }
 
-
-class Compiler(object):
-    def __init__(self, out):
-        self.out = out
-
-    def commas(self, asts, sep=','):
-        if asts:
-            for child in asts[:-1]:
-                self.compile(child)
-                self.out.write(sep)
-            self.compile(asts[-1])
-
-    def compile(self, ast):
-        if ast.tag == 'Program':
-            self.out.write("""\
-/* AUTOMATICALLY GENERATED -- EDIT AT OWN RISK */
+PRELUDE = r"""
+/* AUTOMATICALLY GENERATED -- EDIT AT YOUR OWN RISK */
 
 /*
 var stdin = process.openStdin();
@@ -60,154 +47,163 @@ var equal_tagged_value = function(tv1, tv2)
 {
     return (tv1.tag === tv2.tag) && (tv1.value === tv2.value);
 }
-""")
-            for child in ast.children:
-                self.compile(child)
-            self.out.write("""\
+"""
+
+POSTLUDE = """\
 
 var result = main();
 if (result !== undefined && result !== null)
   print(repr(result));
-""")
+"""
+
+
+class Compiler(BaseCompiler):
+
+    def compile(self, ast):
+        if ast.tag == 'Program':
+            self.write(PRELUDE)
+            for child in ast.children:
+                self.compile(child)
+            self.write(POSTLUDE)
         elif ast.tag == 'Defn':
-            self.out.write('var %s = ' % ast.value)
+            self.write('var %s = ' % ast.value)
             self.compile(ast.children[0])
-            self.out.write(';\n')
+            self.write(';\n')
         elif ast.tag == 'Forward':
             pass
         elif ast.tag == 'StructDefn':
             field_defns = ast.children[0].children
-            self.out.write('function equal_%s(a, b) {\n' % ast.value)
+            self.write('function equal_%s(a, b) {\n' % ast.value)
             for child in field_defns:
                 assert child.tag == 'FieldDefn', child.tag
                 struct_type = child.children[0].value if child.children[0].tag == 'StructType' else None
                 if struct_type:
-                    self.out.write('if (!equal_%s(a.%s, b.%s)) return false;\n' % (struct_type, child.value, child.value))
+                    self.write('if (!equal_%s(a.%s, b.%s)) return false;\n' % (struct_type, child.value, child.value))
                 else:
-                    self.out.write('if (a.%s !== b.%s) return false;\n' % (child.value, child.value))
-            self.out.write('return true;\n')
-            self.out.write('}\n\n')
+                    self.write('if (a.%s !== b.%s) return false;\n' % (child.value, child.value))
+            self.write('return true;\n')
+            self.write('}\n\n')
         elif ast.tag == 'FunLit':
-            self.out.write('function(')
+            self.write('function(')
             self.compile(ast.children[0])
-            self.out.write(')\n')
+            self.write(')\n')
             self.compile(ast.children[1])
         elif ast.tag == 'Args':
             self.commas(ast.children)
         elif ast.tag == 'Arg':
-            self.out.write(ast.value)
+            self.write(ast.value)
         elif ast.tag == 'Body':
-            self.out.write('{')
+            self.write('{')
             self.compile(ast.children[0])
             assert ast.children[1].tag == 'Block'
             block = ast.children[1]
             for child in block.children:
                 self.compile(child)
-                self.out.write(';\n')
-            self.out.write('}')
+                self.write(';\n')
+            self.write('}')
         elif ast.tag == 'VarDecls':
             for child in ast.children:
                 self.compile(child)
         elif ast.tag == 'VarDecl':
-            self.out.write('var %s;\n' % ast.value)
+            self.write('var %s;\n' % ast.value)
         elif ast.tag == 'Block':
-            self.out.write('{')
+            self.write('{')
             for child in ast.children:
                 self.compile(child)
-                self.out.write(';\n')
-            self.out.write('}')
+                self.write(';\n')
+            self.write('}')
         elif ast.tag == 'While':
-            self.out.write('while (')
+            self.write('while (')
             self.compile(ast.children[0])
-            self.out.write(')')
+            self.write(')')
             self.compile(ast.children[1])
         elif ast.tag == 'Op':
             if ast.value == '==' and isinstance(ast.children[0].type, Struct):
-                self.out.write('equal_%s(' % ast.children[0].type.name)
+                self.write('equal_%s(' % ast.children[0].type.name)
                 self.compile(ast.children[0])
-                self.out.write(', ')
+                self.write(', ')
                 self.compile(ast.children[1])
-                self.out.write(')')
+                self.write(')')
             elif ast.value == '==' and isinstance(ast.children[0].type, Union):
-                self.out.write('equal_tagged_value(')
+                self.write('equal_tagged_value(')
                 self.compile(ast.children[0])
-                self.out.write(', ')
+                self.write(', ')
                 self.compile(ast.children[1])
-                self.out.write(')')
+                self.write(')')
             else:
-                self.out.write('(')
+                self.write('(')
                 self.compile(ast.children[0])
-                self.out.write(' %s ' % OPS.get(ast.value, ast.value))
+                self.write(' %s ' % OPS.get(ast.value, ast.value))
                 self.compile(ast.children[1])
-                self.out.write(')')
+                self.write(')')
         elif ast.tag == 'VarRef':
-            self.out.write(ast.value)
+            self.write(ast.value)
         elif ast.tag == 'FunCall':
             self.compile(ast.children[0])
-            self.out.write('(')
+            self.write('(')
             self.commas(ast.children[1:])
-            self.out.write(')')
+            self.write(')')
         elif ast.tag == 'If':
-            self.out.write('if(')
+            self.write('if(')
             self.compile(ast.children[0])
-            self.out.write(')')
+            self.write(')')
             if len(ast.children) == 3:  # if-else
                 self.compile(ast.children[1])
-                self.out.write(' else ')
+                self.write(' else ')
                 self.compile(ast.children[2])
             else:  # just-if
                 self.compile(ast.children[1])
         elif ast.tag == 'Return':
-            self.out.write('return ')
+            self.write('return ')
             self.compile(ast.children[0])
         elif ast.tag == 'Break':
-            self.out.write('break')
+            self.write('break')
         elif ast.tag == 'Not':
-            self.out.write('!(')
+            self.write('!(')
             self.compile(ast.children[0])
-            self.out.write(')')
+            self.write(')')
         elif ast.tag == 'None':
-            self.out.write('null')
+            self.write('null')
         elif ast.tag == 'IntLit':
-            self.out.write(str(ast.value))
+            self.write(str(ast.value))
         elif ast.tag == 'StrLit':
-            self.out.write("'%s'" % ast.value)
+            self.write("'%s'" % ast.value)
         elif ast.tag == 'BoolLit':
             if ast.value:
-                self.out.write("true")
+                self.write("true")
             else:
-                self.out.write("false")
+                self.write("false")
         elif ast.tag == 'Assignment':
             self.compile(ast.children[0])
-            self.out.write(' = ')
+            self.write(' = ')
             self.compile(ast.children[1])
         elif ast.tag == 'Make':
-            self.out.write('{')
+            self.write('{')
             self.commas(ast.children[1:])
-            self.out.write('}')
+            self.write('}')
         elif ast.tag == 'FieldInit':
-            self.out.write("'%s':" % ast.value)
+            self.write("'%s':" % ast.value)
             self.compile(ast.children[0])
         elif ast.tag == 'Index':
             self.compile(ast.children[0])
-            self.out.write('.%s' % ast.value)
+            self.write('.%s' % ast.value)
         elif ast.tag == 'TypeCast':
-            self.out.write("['%s'," % str(ast.children[0].type))
+            self.write("['%s'," % str(ast.children[0].type))
             self.compile(ast.children[0])
-            self.out.write(']')
+            self.write(']')
         elif ast.tag == 'TypeCase':
-            self.out.write('if (')
+            self.write('if (')
             self.compile(ast.children[0])
-            self.out.write("[0] == '%s')" % str(ast.children[1].type))
-            self.out.write('{ var save=')
+            self.write("[0] == '%s')" % str(ast.children[1].type))
+            self.write('{ var save=')
             self.compile(ast.children[0])
-            self.out.write('; ')
+            self.write('; ')
             self.compile(ast.children[0])
-            self.out.write('=')
+            self.write('=')
             self.compile(ast.children[0])
-            self.out.write('[1]; ')
+            self.write('[1]; ')
             self.compile(ast.children[2])
             self.compile(ast.children[0])
-            self.out.write(' =save; }')
+            self.write(' =save; }')
         else:
             raise NotImplementedError(repr(ast))